requirements.txt
alembic==1.18.4
altair==6.0.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
bcrypt==4.1.2
beautifulsoup4==4.14.3
blinker==1.9.0
buildozer==1.5.0
CacheControl==0.14.4
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.4.0
docutils==0.22.4
fastapi==0.133.1
filelock==3.24.2
filetype==1.2.0
firebase-admin==6.4.0
Flask==3.0.0
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
hypothesis==6.98.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pefile==2024.8.26
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
ptyprocess==0.7.0
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
PyJWT==2.11.0
pyparsing==3.3.2
pypiwin32==223
pyserial==3.5
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
referencing==0.37.0
requests==2.32.5
RouterOS-api==0.17.0
rpds-py==0.30.0
rsa==4.9.1
setuptools==82.0.0
sh==2.2.2
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
streamlit==1.54.0
tenacity==9.1.4
tinydb==4.8.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==20.37.0
watchdog==6.0.0
Werkzeug==3.1.5
WTForms==3.2.1
